+++ R/md.R
@@ -67,6 +67,14 @@ to_md <- function(ht, header = TRUE, min_width = getOption("width") / 4, max_wid
   if (!is.na(caption_data$text)) {
     result <- paste0(result, "Table: ", caption_data$text, "\n")
   }
+  notes <- table_notes(ht)
+  if (length(notes) > 0L) {
+    result <- paste0(
+      result,
+      paste0(notes, collapse = "\n\n"),
+      "\n"
+    )
+  }
 
   result
 }